% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changearg.R
\name{changearg}
\alias{changearg}
\title{Change Function Default Arguments}
\usage{
changearg(FUN, ...)
}
\arguments{
\item{FUN}{name of function}

\item{...}{arguments to be overriden (e.g. format = 'member')}
}
\description{
Override default arguments of functions. This functionality is used to
deal with the updated default representation in the SpecsVerification 
package (>= v0.5).
}
