% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyVerification.R
\docType{package}
\name{easyVerification}
\alias{easyVerification}
\alias{easyVerification-package}
\title{Ensemble Forecast Verification for Large Data Sets}
\description{
Set of tools to simplify application of atomic forecast
verification metrics for (comparative) verification of ensemble forecasts
to large data sets. The forecast metrics are imported from the
'SpecsVerification' package, and additional forecast metrics are provided
with this package. Alternatively, new user-defined forecast scores can be
implemented using the example scores provided and applied using the
functionality of this package.
}
