% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EnsIgnss.R
\name{EnsIgnss}
\alias{EnsIgnss}
\title{Compute various ensemble mean error skill scores}
\usage{
EnsIgnss(ens, ens.ref, obs, type = 3)
}
\arguments{
\item{ens}{n x j matrix of n probability forecasts for j categories}

\item{ens.ref}{n x j matrix of n probability forecasts for j categories}

\item{obs}{n x j matrix of occurence of n verifying observations in j categories}

\item{type}{selection of plotting positions to convert ensemble counts to
  probabilities (default to 3, see \code{\link{count2prob}}}
}
\description{
\code{EnsMess} computes the mean error, \code{EnsMaess} the mean absolute error,
\code{EnsMsess} the mean squared error, and \code{EnsRmsess} the square root of
the mean squared error (for consistency with the veri package).
}
\examples{
tm <- toymodel()

## compute skill score
veriApply("EnsIgnss", fcst=tm$fcst, obs=tm$obs, prob=1:3/2)
}
\seealso{
\code{\link{veriApply}}, \code{\link{EnsIgn}}
}

