% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EnsRocss}
\alias{EnsRocss}
\title{Skill score for area under the ROC curve}
\usage{
EnsRocss(ens, ens.ref, obs)
}
\arguments{
\item{ens}{n x j matrix of n probability forecasts for j categories}

\item{ens.ref}{placeholder for consistency with skill scores in SpecsVerification}

\item{obs}{n x j matrix of occurence of n verifying observations in j categories}
}
\description{
Computes the skill score for the area under the ROC curve
}

