% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Size of Array or Vector}
\usage{
size(x)
}
\arguments{
\item{x}{array or vector}
}
\description{
Return dimension of array or length of vector.
}
\examples{
tm <- toymodel()

sapply(tm, size)

}
\seealso{
\code{\link{veriApply}}
}
\keyword{utilities}
