% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EnsCorr}
\alias{EnsCorr}
\title{Ensemble mean correlation}
\usage{
EnsCorr(ens, obs)
}
\arguments{
\item{ens}{n x k matrix of n forecasts from k ensemble members}

\item{obs}{n verifying observations}
}
\description{
Computes the ensemble mean correlation (Pearson) with the veryfing observations
}
\examples{
obs <- rnorm(10)
fcst <- array(rnorm(30), c(10,3)) + obs
EnsCorr(fcst, obs)
}

