% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EnsIgn.R
\name{EnsIgn}
\alias{EnsIgn}
\title{Ignorance score}
\usage{
EnsIgn(ens, obs, type = 3, ...)
}
\arguments{
\item{ens}{n x j matrix of n probability forecasts for j categories}

\item{obs}{n x j matrix of occurence of n verifying observations in j categories}

\item{type}{selection of plotting positions to convert ensemble counts to
  probabilities (default to 3, see \code{\link{count2prob}}}
}
\description{
Computes the ignorance score for an interpretation of the ensemble as a
probability forecast
}
\examples{
tm <- toymodel()

## compute ROC area for tercile forecasts using veriApply
veriApply("EnsIgn", fcst=tm$fcst, obs=tm$obs, prob=1:2/3)
}
\references{
Wilks, D.S. (2011). Statistical methods in the atmospheric sciences (Third Edition).
Academic press.
Jolliffe, I.T. and D.B. Stephenson (2012). Forecast Verification. A Practitioner's Guide
in Atmospheric Science. Wiley-Blackwell.
}
\seealso{
\code{\link{veriApply}}, \code{\link{EnsIgnss}}, \code{\link{count2prob}}
}

