% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsRoca.R
\name{EnsRoca}
\alias{EnsRoca}
\title{Area Under the ROC Curve}
\usage{
EnsRoca(ens, obs)
}
\arguments{
\item{ens}{n x j matrix of n probability forecasts for j categories}

\item{obs}{n x j matrix of occurence of n verifying observations in j categories}
}
\description{
Computes the area under the ROC curve given the observations
}
\examples{
tm <- toymodel()

## compute ROC area for tercile forecasts using veriApply
veriApply("EnsRoca", fcst=tm$fcst, obs=tm$obs, prob=1:2/3)

}
\seealso{
\code{\link{veriApply}}, \code{\link{EnsRocss}}
}

