% Generated by roxygen2 (4.0.2): do not edit by hand
\name{EnsRoca}
\alias{EnsRoca}
\title{Area under the ROC curve}
\usage{
EnsRoca(ens, obs)
}
\arguments{
\item{ens}{n x j matrix of n probability forecasts for j categories}

\item{obs}{n x j matrix of occurence of n verifying observations in j categories}
}
\description{
Computes the area under the ROC curve given the observations
}

