% Generated by roxygen2 (4.0.2): do not edit by hand
\name{veriApply}
\alias{veriApply}
\title{Apply verification metrics from Specs-Verification to large datasets}
\usage{
veriApply(verifun, fcst, obs, tdim = length(dim(fcst)) - 1,
  ensdim = length(dim(fcst)), prob = NULL, threshold = NULL,
  na.rm = FALSE, ...)
}
\arguments{
\item{verifun}{Name of function to compute verification metric (score, skill score)}

\item{fcst}{array of forecast values (at least 2-dimensional)}

\item{obs}{array or vector of verifying observations}

\item{tdim}{index of dimension with the different forecasts}

\item{ensdim}{index of dimension with the different ensemble members}

\item{prob}{probability threshold for category forecasts}

\item{threshold}{absolute threshold for category forecasts}
}
\description{
Apply verification metrics from Specs-Verification to large datasets
}
\keyword{utilities}

