% Generated by roxygen2 (4.0.2): do not edit by hand
\name{veriApply}
\alias{veriApply}
\title{Apply verification metrics to large datasets}
\usage{
veriApply(verifun, fcst, obs, fcst.ref = NULL, tdim = length(dim(fcst)) - 1,
  ensdim = length(dim(fcst)), prob = NULL, threshold = NULL,
  na.rm = FALSE, ...)
}
\arguments{
\item{verifun}{Name of function to compute verification metric (score, skill score)}

\item{fcst}{array of forecast values (at least 2-dimensional)}

\item{obs}{array or vector of verifying observations}

\item{fcst.ref}{array of forecast values for the reference forecast (skill scores only)}

\item{tdim}{index of dimension with the different forecasts}

\item{ensdim}{index of dimension with the different ensemble members}

\item{prob}{probability threshold for category forecasts}

\item{threshold}{absolute threshold for category forecasts}

\item{na.rm}{logical, should incomplete forecasts be used?}

\item{...}{additional arguments passed to \code{verifun}}
}
\description{
Apply verification metrics to large datasets
}
\examples{
obs <- array(rnorm(1000*30), c(1000,30))
fcst <- array(rnorm(1000*30*50), c(1000, 30, 50)) + 0.2*as.vector(obs)
f.me <- veriApply('EnsMe', fcst, obs)
}
\keyword{utilities}

