% Generated by roxygen2 (4.0.2): do not edit by hand
\name{veriUnwrap}
\alias{veriUnwrap}
\title{Unwrap arguments and hand over to function}
\usage{
veriUnwrap(x, verifun, prob = NULL, threshold = NULL, ...)
}
\arguments{
\item{x}{n x k + 1 matrix with n forecasts of k ensemble members
plus the verifying observations}

\item{verifun}{character string with function name to be executed}

\item{prob}{probability threshold for conversion of forecasts to
probability forecasts as in \code{\link{convert2prob}}}

\item{threshold}{absolute threshold for conversion of forecasts to
probability forecasts as in \code{\link{convert2prob}}}

\item{...}{additional arguments passed on to \code{verifun}}
}
\description{
decomposes input argument into forecast and verifying observations
and hands these over to the function provided
}
\details{
Only forecasts with non-missing observation and complete ensembles are
computed. All other forecasts are set to missing. For aggregate metrics
(e.g. skill scores) the metric is computed over non-missing observation/forecast
pairs only.
}

