% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veriUnwrap.R
\name{veriUnwrap}
\alias{veriUnwrap}
\title{Unwrap Arguments and Hand Over to Verification Function}
\usage{
veriUnwrap(x, verifun, nind = c(nens = ncol(x) - 1, nref = 0, nobs = 1, nprob
  = 0, nthresh = 0), ref.opts = NULL, ...)
}
\arguments{
\item{x}{n x k + 1 matrix with n forecasts of k ensemble members plus the 
verifying observations}

\item{verifun}{character string with function name to be executed}

\item{nind}{named vector with number of ensemble members, ensemble members of
reference forecasts, observations (defaults to 1), probability or absolute 
thresholds (see details)}

\item{ref.opts}{string or list with specifications for the reference forecast
(see details)}

\item{...}{additional arguments passed on to \code{verifun}}
}
\description{
Decomposes input arguments into forecast, verifying observations, and 
reference forecast and hands these over to the function provided.
}
\details{
Forecast verification metrics are only computed for forecasts with 
  non-missing verifying observation and at least one non-missing ensemble 
  member. Metrics for all other forecasts are set to missing. For aggregate 
  metrics (e.g. skill scores) the metric is computed over non-missing 
  observation/forecast pairs only.
  
  For computation of skill scores, reference forecasts can be provided. That 
  is, the first \code{nens} columns of \code{x} contain the forecasts, the 
  \code{(nens + 1):(ncol(x) - 1)} following columns contain the reference 
  forecast, and the final column contains the observations. If no reference 
  forecast is provided (i.e. \code{ncol(x) == nens + 1}), a climatological 
  forecast is constructed from the \code{n} verifying observations.
  
  The elements of vector \code{nind} have to be named with \code{nens} 
  containing the number of ensemble members, \code{nref} the number of 
  ensemble members in the reference forecast for skill scores, \code{nobs} 
  the number of observations (only one supported), \code{nprob} the number of
  probability thresholds, and \code{nthresh} the number of absolute threshold
  for conversion of continuous forecasts to category forecasts.
  
  \code{ref.opts} specifies the set-up of the climatological reference 
  forecast for skill scores if no explicit reference forecast is provided. 
  This can be either \code{NULL}, all available observations are used as 
  equiprobable members of a reference forecast, or a list with \code{n} 
  elements containing the indices of the observations to be used to construct
  the reference forecast for forecast \code{n}. The indices provided have to 
  be non-missing and in the range of \code{1} to \code{n} of the verifying 
  observations.
}
\note{
Out-of-sample reference forecasts are not fully supported for 
  categorical forecasts defined on the distribution of forecast values (e.g. 
  using the argument \code{prob}). Whereas only the years specified in 
  \code{ref.opts} are used for the reference forecasts, the probability 
  thresholds for the reference forecasts are defined on the collection of
  years specified in \code{ref.opts}.
}
\seealso{
\code{\link{veriApply}}
}

