% Generated by roxygen2 (4.0.2): do not edit by hand
\name{veriUnwrap}
\alias{veriUnwrap}
\title{Unwrap arguments and hand over to function}
\usage{
veriUnwrap(x, verifun, nens = ncol(x) - 1, prob = NULL, threshold = NULL,
  ...)
}
\arguments{
\item{x}{n x k + 1 matrix with n forecasts of k ensemble members
plus the verifying observations}

\item{verifun}{character string with function name to be executed}

\item{nens}{number of ensemble members in forecast (see details)}

\item{prob}{probability threshold for conversion of forecasts to
probability forecasts as in \code{\link{convert2prob}}}

\item{threshold}{absolute threshold for conversion of forecasts to
probability forecasts as in \code{\link{convert2prob}}}

\item{...}{additional arguments passed on to \code{verifun}}
}
\description{
decomposes input argument into forecast and verifying observations
and hands these over to the function provided
}
\details{
Only forecasts with non-missing observation and complete ensembles are
computed. All other forecasts are set to missing. For aggregate metrics
(e.g. skill scores) the metric is computed over non-missing observation/forecast
pairs only.

For computation of skill scores, reference forecasts can be provided. That is, the
first \code{nens} columns of \code{x} contain the forecasts, the \code{(nens + 1):(ncol(x) - 1)}
following columns contain the reference forecast, and the final column contains the observations.
If no reference forecast is provided (i.e. \code{ncol(x) == nens + 1}), a climatological
forecast is constructed from the \code{n} verifying observations.
}

